Fix gfx polyfillCC in native
